Script

+++ src/bouncing_pellet.hpp
@@ -5,8 +5,8 @@
 
 class BouncingPellet : public Entity {
 public:
-    BouncingPellet(float start_x, float start_y, ContinuousAngle angle);
-    void Draw(const DrawPosition &grid_tl_pos, float grid_cell_width, float grid_cell_height) override;
+    BouncingPellet(double start_x, double start_y, ContinuousAngle angle);
+    void Draw(const DrawPosition &grid_tl_pos, double grid_cell_width, double grid_cell_height) override;
     void Update(int fps) override;
 };
